requirements.txt
Django==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
pillow==10.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1